requirements.txt
discord.py
pytz
requests
django
pillow
cairosvg
